refactored include_array_to_hash to a utils module - this also simplified the test for it. yay
