+++ lib/active_model_serializers.rb
@@ -15,6 +15,7 @@ require 'active_model'
 require 'action_controller'
 
 require 'active_model/serializer'
+require 'active_model/serializer/utils'
 require 'active_model/serializable_resource'
 require 'active_model/serializer/version'
